integrate RenderGraph version into existing RendererFeature and fixed mipmap creation
