--- Runtime/Scripts/RenderPipelines/RoughRefractionPass.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 0606be9544304c7a94b493c4c7804bee
-timeCreated: 1710326284
